+++ demo/app/pages/+default/home/home.component.ts
@@ -1,7 +1,7 @@
 import {Component, ChangeDetectionStrategy} from '@angular/core';
 import {ComponentRedirectRoute, ComponentRoute} from '@anglr/common/router';
 import {WithScrollableCssClass} from '@anglr/common';
-import { Select, Option, OptionGroup, OptionGroupTemplate, SelectAbsolute } from '@anglr/select';
+import { Select, Option, SelectAbsolute } from '@anglr/select';
 
 /**
  * Home component
@@ -14,8 +14,6 @@ import { Select, Option, OptionGroup, OptionGroupTemplate, SelectAbsolute } from
     [
         Select,
         Option,
-        OptionGroup,
-        OptionGroupTemplate,
         SelectAbsolute,
     ],
     changeDetection: ChangeDetectionStrategy.OnPush