+++ src/plugins/popup/simplePopup/simplePopup.component.ts
@@ -1,5 +1,6 @@
-import {ChangeDetectionStrategy, Component, ElementRef, Inject, inject, Optional} from '@angular/core';
+import {ChangeDetectionStrategy, Component, effect, ElementRef, Inject, inject, Optional, Signal, viewChild} from '@angular/core';
 import {NgTemplateOutlet} from '@angular/common';
+import {LocalizePipe} from '@anglr/common';
 import {RecursivePartial} from '@jscrpt/common';
 import {deepCopyWithArrayOverride} from '@jscrpt/common/lodash';
 
@@ -7,11 +8,14 @@ import {Popup, PopupOptions, SelectOptionState, SelectPlugin, SimplePopupCssClas
 import {SelectPluginInstances, SelectBus} from '../../../misc/classes';
 import {CopyOptionsAsSignal} from '../../../decorators';
 import {POPUP_OPTIONS} from '../../../misc/tokens';
-import {GroupedListOptions} from '../../../pipes';
+import {DisplayValue, GroupedListOptions} from '../../../pipes';
+
+//TODO: hover into mouseover => use "active"
 
 const defaultOptions: PopupOptions<SimplePopupCssClasses> =
 {
     visible: false,
+    liveSearchEnabled: true,
     cssClasses:
     {
         componentElement: 'popup-component',
@@ -40,6 +44,8 @@ const defaultOptions: PopupOptions<SimplePopupCssClasses> =
     },
     imports:
     [
+        LocalizePipe,
+        DisplayValue,
         NgTemplateOutlet,
         GroupedListOptions,
     ],
@@ -70,11 +76,33 @@ export class SimplePopup<TValue = unknown> implements Popup<TValue, PopupOptions
      */
     public selectBus: SelectBus<TValue> = inject(SelectBus) as SelectBus<TValue>;
 
+    //######################### protected properties - children #########################
+
+    /**
+     * Instance of element that is used as placeholder for live search
+     */
+    protected liveSearchPlaceholder: Signal<ElementRef<HTMLElement>|undefined|null> = viewChild('liveSearchPlaceholder', {read: ElementRef});
+
     //######################### constructor #########################
     constructor(@Inject(POPUP_OPTIONS) @Optional() options?: RecursivePartial<PopupOptions<SimplePopupCssClasses>>|null,)
     {
         this.options = deepCopyWithArrayOverride(defaultOptions as PopupOptions<SimplePopupCssClasses>,
                                                  options);
+
+        effect(() =>
+        {
+            if(this.options.liveSearchEnabled)
+            {
+                const placeholder = this.liveSearchPlaceholder();
+
+                if(!placeholder)
+                {
+                    return;
+                }
+
+                placeholder.nativeElement.after(this.selectPlugins.LiveSearch.pluginElement.nativeElement);
+            }
+        });
     }
 
     //######################### protected methods - template bindings #########################